# feat(perf): use Mmap in `ManyCar::read_only_files`