--- a/Cargo.toml
+++ b/Cargo.toml
@@ -111,6 +111,7 @@ libp2p = { version = "0.52.1", default-features = false, features = [
 ] }
 libsecp256k1 = "0.7"
 lru = "0.11"
+memmap2 = "0.7"
 memory-stats = "1.1"
 mimalloc = { version = "0.1.34", optional = true, default-features = false }
 nom = "7.1.3"