## Summary of changes

<!-- Please write a comprehensive summary of your changes and what was the motivation behind them -->

Related to https://github.com/ChainSafe/forest/issues/3361

Changes introduced in this pull request:

- use Mmap in `ManyCar::read_only_files` for better performance.

P.S. benchmark shows the performance of using vanilla Mmap API is 2-5x better than using a Mmap wrapper that implements `positioned-io` traits, we could benchmark and discuss about it later

### `forest-tool benchmark graph-traversal` result

calibnet snapshot(local laptop):
```console
main branch:
traversed 2.84 GiB at 104.91 MiB/s in 00:00:27
traversed 2.84 GiB at 104.65 MiB/s in 00:00:27
traversed 2.84 GiB at 103.04 MiB/s in 00:00:28

feature branch:
traversed 2.84 GiB at 139.04 MiB/s in 00:00:20
traversed 2.84 GiB at 157.26 MiB/s in 00:00:18
traversed 2.84 GiB at 161.50 MiB/s in 00:00:17
```

mainnet snapshot(DO droplet):
```console
# first run
➜  car_db /root/forest-tool-main benchmark graph-traversal 3134862.forest.car.zst
  traversed 93.80 GiB at 25.01 MiB/s in 01:04:01
➜  car_db /root/forest-tool-mmap benchmark graph-traversal 3134862.forest.car.zst
  traversed 93.80 GiB at 95.05 MiB/s in 00:16:50
# second run
➜  car_db /root/forest-tool-main benchmark graph-traversal 3134862.forest.car.zst
  traversed 93.80 GiB at 36.10 MiB/s in 00:44:20
➜  car_db /root/forest-tool-mmap benchmark graph-traversal 3134862.forest.car.zst
  traversed 93.80 GiB at 93.92 MiB/s in 00:17:02
```

### `forest-tool snapshot validate [snapshot].forest.car.zst` result (with hyperfine)

calibnet(local laptop):
```
# hyperfine -w 1 "./target/release/forest-tool-mmap snapshot validate ~/fr/calibnet/paritydb/car_db/1692774391.forest.car.zst" "./target/release/forest-tool-main snapshot validate ~/fr/calibnet/paritydb/car_db/1692774391.forest.car.zst"

Benchmark 1: ./target/release/forest-tool-mmap snapshot validate ~/fr/calibnet/paritydb/car_db/1692774391.forest.car.zst
  Time (mean ± σ):     14.822 s ±  2.366 s    [User: 16.257 s, System: 0.449 s]
  Range (min … max):    9.717 s … 16.555 s    10 runs

Benchmark 2: ./target/release/forest-tool-main snapshot validate ~/fr/calibnet/paritydb/car_db/1692774391.forest.car.zst
  Time (mean ± σ):     16.309 s ±  2.409 s    [User: 16.436 s, System: 1.565 s]
  Range (min … max):    9.617 s … 18.128 s    10 runs

Summary
  './target/release/forest-tool-mmap snapshot validate ~/fr/calibnet/paritydb/car_db/1692774391.forest.car.zst' ran
    1.10 ± 0.24 times faster than './target/release/forest-tool-main snapshot validate ~/fr/calibnet/paritydb/car_db/1692774391.forest.car.zst'
```

mainnet(DO droplet):
```
# hyperfine -w 0 -r 3 "/root/forest-tool-mmap snapshot validate 3134862.forest.car.zst" "/root/forest-tool-main snapshot validate 3134862.forest.car.zst"

Benchmark 1: /root/forest-tool-mmap snapshot validate 3134862.forest.car.zst
  Time (mean ± σ):     1773.404 s ± 13.463 s    [User: 2020.588 s, System: 59.360 s]
  Range (min … max):   1765.198 s … 1788.942 s    3 runs

Benchmark 2: /root/forest-tool-main snapshot validate 3134862.forest.car.zst
  Time (mean ± σ):     2767.611 s ± 286.282 s    [User: 2123.710 s, System: 365.960 s]
  Range (min … max):   2444.817 s … 2990.734 s    3 runs

Summary
  /root/forest-tool-mmap snapshot validate 3134862.forest.car.zst ran
    1.56 ± 0.16 times faster than /root/forest-tool-main snapshot validate 3134862.forest.car.zst
```

## Reference issue to close (if applicable)

<!-- Include the issue reference this pull request is connected to -->
<!-- See more keywords here https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword -->
<!--(e.g. Closes #1)-->

Closes

## Other information and links

<!-- Add any other context about the pull request here. Those might be helpful links based on your investigation, relevant commits from this or other repositories or anything else -->

## Change checklist

<!-- Please add a changelog entry for your change if needed. -->
<!-- Follow this format https://keepachangelog.com/en/1.0.0/ -->

- [x] I have performed a self-review of my own code,
- [x] I have made corresponding changes to the documentation,
- [x] I have added tests that prove my fix is effective or that my feature works (if possible),
- [x] I have made sure the [CHANGELOG][1] is up-to-date. All user-facing changes should be reflected in this document.

<!-- Thank you 🔥 -->

[1]: https://github.com/ChainSafe/forest/blob/main/CHANGELOG.md