@lemmih 2023-08-24
Let's use an environment variable to control whether to use file-io or mmap. And let's benchmark tipset validation. You can use hyperfine for this.