@hanabi1224 2023-08-24
> Let's use an environment variable to control whether to use file-io or mmap. And let's benchmark tipset validation. You can use hyperfine for this.

@lemmih Just to confirm, do you mean `forest-tool snapshot validate [foo].forest.car.zst` with default CLI options?