@hanabi1224 2023-08-24
@lemmih `forest-tool snapshot validate` benchmark result updated in the PR description. Btw, which approach should be made default when the environment variable is not set?  Should I use mmap as default with `FOREST_CAR_LOADER_FILE_IO=1` as the switch or use file-io as default with `FOREST_CAR_LOADER_MMAP=1` as the switch?